requirements.txt
Babel==2.12.1
blinker==1.6.2
cachelib==0.9.0
click==8.1.7
colorama==0.4.6
dnspython==2.4.2
email-validator==2.0.0.post2
Flask==2.2.5
flask-babel==3.1.0
Flask-Caching==2.0.2
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-Moment==1.0.5
flask-mongoengine==1.0.0
Flask-WTF==1.1.1
gunicorn==21.2.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mongoengine==0.27.0
packaging==23.1
Pillow==10.0.0
PyJWT==2.8.0
pymongo==4.4.1
python-dotenv==1.0.0
python-slugify==8.0.1
pytz==2023.3
slugify==0.0.1
text-unidecode==1.3
Unidecode==1.3.6
Werkzeug==2.3.7
WTForms==3.0.1
